+++ CHANGELOG.md
@@ -1,4 +1,19 @@
-# swift-docker changelog
+# SwiftDocker Changelog
+
+## 2016-11-27 • Version 2
+
+This version focuses on improving the Swift build automation script.
+
+The following changes have been made:
+
+- The Swift script has been converted to a Swift Package
+- Improved manifest parsing (100% safer!)
+- Better Console I/O
+
+Other changes have been made:
+
+- The Dockerfiles are no longer committed, since they don't provide any added value
+- Fixes in the Dockerfile (permissions fixed, and extraneous /sta directory removed)
 
 ## 2016-11-10 • Version 1